Fix Bug 1129624: Refactor get_download_url
