+++ bedrock/firefox/templates/firefox/android/faq.html
@@ -24,7 +24,7 @@
 {% block content %}
 
 <div id="main-feature">
-  {{ download_firefox(mobile=True, small=True) }}
+  {{ download_firefox(platform='android', small=True) }}
   <h1 class="large">{{_('Mobile FAQ')}}</h1>
 </div>
