+++ bedrock/firefox/templates/firefox/android/index.html
@@ -353,7 +353,7 @@
               <a rel="external" href="https://support.mozilla.org/kb/will-firefox-work-my-mobile-device">{{ _('Supported devices') }}</a>
             </li>
             <li>
-              <a href="{{ product_url('mobile', 'notes') }}">{{ _('Release notes') }}</a>
+              <a href="{{ firefox_url('android', 'notes') }}">{{ _('Release notes') }}</a>
             </li>
           </ul>
         </div>{#-- /#download-content --#}
