+++ bedrock/firefox/templates/firefox/channel.html
@@ -48,10 +48,10 @@
       {# L10n: This description applies to Firefox Beta #}
       <h3>{{_('The latest features in a more stable environment')}}</h3>
       <div class="download-box" id="beta-desktop">
-        {{ download_firefox('beta', icon=False, small=True, force_direct=True) }}
+        {{ download_firefox('beta', icon=False, platform='desktop', small=True, force_direct=True) }}
       </div>
-      <div class="download-box mobile" id="beta-mobile">
-        {{ download_firefox('beta', icon=False, mobile=True, small=True) }}
+      <div class="download-box android" id="beta-android">
+        {{ download_firefox('beta', icon=False, platform='android', small=True) }}
       </div>
       {% if l10n_has_tag('telemetry_notice') %}
         <p class="warning">
@@ -72,10 +72,10 @@
       {# L10n: This description applies to Firefox on the Release channel #}
       <h3>{{_('Tried, tested and used by millions around the world')}}</h3>
       <div class="download-box" id="firefox-desktop">
-        {{ download_firefox(icon=False, small=True, force_direct=True) }}
+        {{ download_firefox(icon=False, platform='desktop', small=True, force_direct=True) }}
       </div>
-      <div class="download-box mobile" id="firefox-mobile">
-        {{ download_firefox(icon=False, mobile=True, small=True) }}
+      <div class="download-box android" id="firefox-android">
+        {{ download_firefox(icon=False, platform='android', small=True) }}
       </div>
       <p class="more"><a href="{{ php_url('/firefox/') }}">{{_('Learn more about Firefox')}}</a></p>
     </div>
@@ -89,10 +89,10 @@
         <h3>{{_('The newest innovations in an experimental environment')}}</h3>
       {% endif %}
       <div class="download-box" id="aurora-desktop">
-        {{ download_firefox('aurora', icon=False, small=True) }}
+        {{ download_firefox('alpha', icon=False, platform='desktop', small=True) }}
       </div>
-      <div class="download-box mobile" id="aurora-mobile">
-        {{ download_firefox('aurora', icon=False, mobile=True, small=True) }}
+      <div class="download-box android" id="aurora-android">
+        {{ download_firefox('alpha', icon=False, platform='android', small=True) }}
       </div>
       <p class="warning">
         {% if l10n_has_tag('channel_dev_edition') %}
