+++ bedrock/firefox/templates/firefox/firstrun.html
@@ -53,7 +53,7 @@
 <section id="colophon" class="billboard">
   <nav>
     <ul class="primary">
-      <li> <a href="{{ product_url('firefox', 'notes') }}">{{ _('Release Notes') }}</a></li>
+      <li> <a href="{{ firefox_url('desktop', 'notes') }}">{{ _('Release Notes') }}</a></li>
       <li><a href="{{ url('firefox.desktop.index') }}">{{ _('Firefox Features') }}</a></li>
       <li><a href="https://support.mozilla.org/">{{ _('Firefox Help') }}</a></li>
     </ul>
