+++ bedrock/firefox/templates/firefox/installer-help.html
@@ -67,7 +67,7 @@
       </div>
 
       <div class="download-box" id="aurora-desktop">
-        {{ download_firefox('aurora', small=True, force_direct=True, force_full_installer=True, icon=False, locale=installer_lang) }}
+        {{ download_firefox('alpha', small=True, force_direct=True, force_full_installer=True, icon=False, locale=installer_lang) }}
       </div>
     {% endif %}
   </div>
