+++ bedrock/firefox/templates/firefox/organizations/faq.html
@@ -45,7 +45,7 @@
 </p>
 
 <h4>{{ _('Where can I download Mozilla Firefox ESR?') }}</h4>
-<p>{{ _('You can <a href="%s">download Mozilla Firefox ESR here</a>.')|format(product_url('firefox', 'all', 'esr')) }}</p>
+<p>{{ _('You can <a href="%s">download Mozilla Firefox ESR here</a>.')|format(firefox_url('desktop', 'all', 'esr')) }}</p>
 
 <h4>{{ _('What does the Mozilla Firefox ESR life cycle look like?') }}</h4>
 <p>
