+++ bedrock/firefox/templates/firefox/organizations/organizations.html
@@ -62,7 +62,7 @@
   <aside id="sidebar" class="sidebar">
 
     <div id="download">
-      <p><a class="button" href="{{ product_url('firefox', 'all', 'esr') }}">{{ _('Download Mozilla Firefox ESR') }}</a></p>
+      <p><a class="button" href="{{ firefox_url('desktop', 'all', 'esr') }}">{{ _('Download Mozilla Firefox ESR') }}</a></p>
     </div>
 
     <div id="alternate-download">
