+++ bedrock/firefox/templates/firefox/sync-old.html
@@ -71,7 +71,7 @@
       </p>
     </section>
     <aside role="complimentary">
-      {{ download_firefox(force_direct=True, mobile=True, small=True) }}
+      {{ download_firefox(force_direct=True, platform='android', small=True) }}
     </aside>
   </article>
 </div>
