+++ bedrock/mozorg/helpers/__init__.py
@@ -1,4 +1,3 @@
 # flake8: noqa
-import download_buttons
 import misc
 import social_widgets
