+++ bedrock/mozorg/helpers/misc.py
@@ -471,44 +471,6 @@ def absolute_url(url):
 
 
 @jingo.register.function
-def product_url(product, page, channel=None):
-    """
-    Return a product-related URL like /firefox/all/ or /mobile/beta/notes/.
-
-    Examples
-    ========
-
-    In Template
-    -----------
-
-        {{ product_url('firefox', 'all', 'organizations') }}
-        {{ product_url('firefox', 'sysreq', channel) }}
-        {{ product_url('mobile', 'notes') }}
-    """
-
-    app = product
-    kwargs = {}
-
-    if product == 'mobile':
-        app = 'firefox'
-
-    # Tweak the channel name for the naming URL pattern in urls.py
-    if channel == 'release':
-        channel = None
-    if channel == 'aurora' and product == 'firefox':
-        channel = 'developer'
-    if channel == 'esr':
-        channel = 'organizations'
-
-    if channel:
-        kwargs['channel'] = channel
-    if page == 'notes':
-        kwargs['product'] = product
-
-    return reverse('%s.%s' % (app, page), kwargs=kwargs)
-
-
-@jingo.register.function
 def releasenotes_url(release):
     prefix = 'aurora' if release.channel == 'Aurora' else 'release'
     if release.product == 'Firefox for Android':
