--- bedrock/mozorg/templates/mozorg/download_buttons/mobile_small_landing.html
@@ -1,19 +1,0 @@
-{# This Source Code Form is subject to the terms of the Mozilla Public
- # License, v. 2.0. If a copy of the MPL was not distributed with this
- # file, You can obtain one at http://mozilla.org/MPL/2.0/. -#}
-
-{% extends "mozorg/download_buttons/mobile_base.html" %}
-
-{% block class %}download-button-mobile-small{% endblock %}
-
-{% block title %}{{ _('Get Firefox for Android »') }}
-  <span class="download-info">
-    {{ _('Free from the Google Play Store') }}
-  </span>
-{% endblock %}
-
-{% block subtitle %}{% endblock %}
-
-{% block small_print %}{% endblock %}
-
-{% block download_info %}{% endblock %}
