+++ bedrock/mozorg/tests/test_helper_misc.py
@@ -537,75 +537,6 @@ class TestAbsoluteURLFilter(TestCase):
         eq_(absolute_url('https://www.mozilla.org/en-US/firefox/new/'), expected)
 
 
-class TestProductURL(TestCase):
-    rf = RequestFactory()
-
-    def _render(self, product, page, channel=None):
-        req = self.rf.get('/')
-        req.locale = 'en-US'
-        if channel:
-            tmpl = "{{ product_url('%s', '%s', '%s') }}" % (product, page, channel)
-        else:
-            tmpl = "{{ product_url('%s', '%s') }}" % (product, page)
-        return render(tmpl, {'request': req})
-
-    def test_firefox_all(self):
-        """Should return a reversed path for the Firefox download page"""
-        eq_(self._render('firefox', 'all'),
-            '/en-US/firefox/all/')
-        eq_(self._render('firefox', 'all', 'release'),
-            '/en-US/firefox/all/')
-        eq_(self._render('firefox', 'all', 'beta'),
-            '/en-US/firefox/beta/all/')
-        eq_(self._render('firefox', 'all', 'aurora'),
-            '/en-US/firefox/developer/all/')
-        eq_(self._render('firefox', 'all', 'esr'),
-            '/en-US/firefox/organizations/all/')
-        eq_(self._render('firefox', 'all', 'organizations'),
-            '/en-US/firefox/organizations/all/')
-
-    def test_firefox_sysreq(self):
-        """Should return a reversed path for the Firefox sysreq page"""
-        eq_(self._render('firefox', 'sysreq'),
-            '/en-US/firefox/system-requirements/')
-        eq_(self._render('firefox', 'sysreq', 'release'),
-            '/en-US/firefox/system-requirements/')
-        eq_(self._render('firefox', 'sysreq', 'beta'),
-            '/en-US/firefox/beta/system-requirements/')
-        eq_(self._render('firefox', 'sysreq', 'aurora'),
-            '/en-US/firefox/developer/system-requirements/')
-        eq_(self._render('firefox', 'sysreq', 'esr'),
-            '/en-US/firefox/organizations/system-requirements/')
-        eq_(self._render('firefox', 'sysreq', 'organizations'),
-            '/en-US/firefox/organizations/system-requirements/')
-
-    def test_firefox_notes(self):
-        """Should return a reversed path for the Firefox notes page"""
-        eq_(self._render('firefox', 'notes'),
-            '/en-US/firefox/notes/')
-        eq_(self._render('firefox', 'notes', 'release'),
-            '/en-US/firefox/notes/')
-        eq_(self._render('firefox', 'notes', 'beta'),
-            '/en-US/firefox/beta/notes/')
-        eq_(self._render('firefox', 'notes', 'aurora'),
-            '/en-US/firefox/developer/notes/')
-        eq_(self._render('firefox', 'notes', 'esr'),
-            '/en-US/firefox/organizations/notes/')
-        eq_(self._render('firefox', 'notes', 'organizations'),
-            '/en-US/firefox/organizations/notes/')
-
-    def test_mobile_notes(self):
-        """Should return a reversed path for the mobile notes page"""
-        eq_(self._render('mobile', 'notes'),
-            '/en-US/mobile/notes/')
-        eq_(self._render('mobile', 'notes', 'release'),
-            '/en-US/mobile/notes/')
-        eq_(self._render('mobile', 'notes', 'beta'),
-            '/en-US/mobile/beta/notes/')
-        eq_(self._render('mobile', 'notes', 'aurora'),
-            '/en-US/mobile/aurora/notes/')
-
-
 class TestReleaseNotesURL(TestCase):
     @patch('bedrock.mozorg.helpers.misc.reverse')
     def test_aurora_android_releasenotes_url(self, mock_reverse):
