+++ media/css/firefox/all.less
@@ -106,7 +106,7 @@
                 background-repeat: no-repeat;
                 background-image: url(/media/img/firefox/all/download-icons.png);
             }
-            &.windows64 a {
+            &.win64 a {
                 background-position: -300px 50%
             }
             &.linux a {
