+++ media/css/firefox/channel.less
@@ -145,7 +145,7 @@
     text-align: left;
 }
 
-.android .download-box.mobile {
+.android .download-box.android {
     display: none;
 }
