+++ media/css/firefox/template-resp.less
@@ -9,7 +9,7 @@
 }
 
 #main-feature .download-button-small,
-#main-feature .download-button-mobile-small {
+#main-feature .download-button-android-small {
     position: absolute;
     top: 0;
     right: 30px;
@@ -28,7 +28,7 @@
 @media only screen and (max-width: @breakTablet) {
 
 #main-feature .download-button-small,
-#main-feature .download-button-mobile-small {
+#main-feature .download-button-android-small {
     position: static;
 }
