+++ media/css/firefox/template.less
@@ -9,7 +9,7 @@
 }
 
 #main-feature .download-button-small,
-#main-feature .download-button-mobile-small {
+#main-feature .download-button-android-small {
     position: absolute;
     top: 0;
     right: 48px;
